#10 サービスファイルがあるか確認する処理の追加

+++ install.sh
@@ -6,19 +6,22 @@
 
 Ver="1.02"
 SERVICE_NAME="dipper.service"
+User_servce="/etc/systemd/system/$SERVICE_NAME"
 
-# サービスの状態を確認
-status=$(systemctl is-active "$SERVICE_NAME")
-# 状態に応じて処理を分岐
-if [ "$status" = "active" ]; then
-    systemctl stop "$SERVICE_NAME"
-fi
+if [ -e ${User_servce} ]; then
+    # サービスの状態を確認
+    status=$(systemctl is-active "$SERVICE_NAME")
+    # 状態に応じて処理を分岐
+    if [ "$status" = "active" ]; then
+        systemctl stop "$SERVICE_NAME"
+    fi
 
-# サービスの有効化状態を確認
-enabled=$(systemctl is-enabled "$SERVICE_NAME")
-# 有効化状態に応じて処理を分岐
-if [ "$enabled" = "enabled" ]; then
-    systemctl disable "$SERVICE_NAME"
+    # サービスの有効化状態を確認
+    enabled=$(systemctl is-enabled "$SERVICE_NAME")
+    # 有効化状態に応じて処理を分岐
+    if [ "$enabled" = "enabled" ]; then
+        systemctl disable "$SERVICE_NAME"
+    fi
 fi
 
 # 以前のバージョンのアンインストール処理
